+++ packages/ui/src/components/Modal/Modal.tsx
@@ -2,7 +2,7 @@ import { forwardRef } from "react";
 import clsx from "clsx";
 import { useDialog } from "@synerity/headless";
 
-import * as styles from "./Modal.module.css";
+import styles from "./Modal.module.css";
 
 export type ModalSize = "sm" | "md" | "lg" | "xl" | "full";