+++ packages/ui/src/components/Spinner/Spinner.tsx
@@ -1,6 +1,6 @@
 import clsx from "clsx";
 
-import * as styles from "./Spinner.module.css";
+import styles from "./Spinner.module.css";
 
 export type SpinnerSize = "xs" | "sm" | "md" | "lg" | "xl";