+++ apps/sample/CHANGELOG.md
@@ -1,5 +1,12 @@
 # sample
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @synerity/ui@0.1.2
+
 ## 0.0.1
 
 ### Patch Changes